Emit graph edges for read/write.
Do some post-processing to match-up reads and writes in the graph.
In the future, we may want to be able to pause processes in order
if they "get ahead of themselves" while we follow a graph. Say a
write -> read occurs, and we see the read return before we see the
write return (could happen!). Then we may want to pause the tracee
that reads until we see the write in order to match it up?